requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.0
requests==2.31.0
playwright==1.41.0
pdfplumber==0.10.4
openai==1.10.0
python-dotenv==1.0.0
pandas==2.1.4
pillow==10.2.0
numpy==1.26.3
matplotlib==3.8.2
beautifulsoup4==4.12.3
aiohttp==3.9.1